Track separate precision defaults for each kind of sampler, give initial defaults as per spec. Also make fragment floats have no default.  Modify/add tests to adapt to these changes.
git-svn-id: https://cvs.khronos.org/svn/repos/ogl/trunk/ecosystem/public/sdk/tools/glslang@22066 e7fa87d3-cd2b-0410-9028-fcbf551c1848

+++ Test/300operations.frag
@@ -1,7 +1,7 @@
 #version 300 es
 
 uniform block {
-    float f;
+    mediump float f;
 } instanceName;
 
 struct S {
